feat: 스코프 설정과 GitHub Source · GitLab Target
스코프 설정(`source.repos`, `since`)은 E2E 를 위해 넣은 비계가 아니다. 조직 전체를
가리키고 10 년치 이슈를 다 받겠다고 결정하는 것은 대개 의도한 바가 아니고,
`since` 는 첫 실행을 가두고 커서를 잃었을 때 피해를 한정한다. 커서 자체는 마지막
성공 실행이 남긴 것이지 이 값이 아니다.

GitHub Source:

  - 저장소 열거는 페이지가 비면 멈춘다. 필터링 후 결과가 아니라 페이지다 —
    archived 저장소도 함께 온다.
  - 이슈 목록이 PR 도 같이 준다(번호 공간이 하나다). `Since` 가 증분을 만든다.
  - 브랜치는 이슈 목록에 없다. MR 은 브랜치가 있어야 만들어지므로 **열린 PR
    에서만** 추가 왕복을 한 번 쓴다. 닫힌 PR 은 어차피 이슈가 된다.

GitLab Target:

  - FindByOrigin 이 매핑을 target 에서 읽어온다. GitLab 의 description 검색이
    fuzzy_search → `ILIKE '%…%'` 한 번이고 쿼리를 단어로 쪼개지 않는다는 것을
    실제 인스턴스 코드로 확인했다. 공백과 구두점이 든 마커가 그대로 매칭된다.
    인덱스 없는 스캔이라 캐시는 둘 수 있지만 target 에서 재구축 가능해야 한다.
  - 검색은 좁히기만 하고 판단은 Contains 가 한다. MR 목록이 `in` 을 안 받아서
    범위가 넓어져도 비용이지 오류가 아니다.
  - EnsureRepo·EnsureLabels 는 이미 있는 경우를 정상 경로로 다룬다. 이 프로그램은
    언제든 중단되고 다시 도는 CronJob 이라 "지난번에 하다 말았다" 가 복구가 아니라
    평상시다.

SDK 는 각 포지의 공식 것을 쓴다(go-github v91, gitlab-org/api/client-go).
git-pkgs/forge 같은 통합 추상화도 있지만 7 개월 된 라이브러리고, 우리가 정작
필요한 쓰기·증분·첨부 지원이 불명확해서 어려운 부분은 어차피 직접 짜면서
추상화와 싸우기까지 할 위험이 있다. 우리 Source/Target 인터페이스를 얇게 두는
쪽이 통제 가능하다.

Ref.ID 를 int64 로 넓혔다. target 이 정한 식별자고 target 마다 폭이 다르다.
Issue.Number 는 source 가 주는 그대로 둔다.

+++ forge/forge.go
@@ -99,5 +99,9 @@ type Ref struct {
 	Kind Kind
 
 	// ID is the target's own number for it — GitLab's iid, for instance.
-	ID int
+	//
+	// Wider than the source's Number on purpose: this is whatever the target
+	// chose, and targets do not agree on the width. Number is the source's and
+	// stays as the source gives it.
+	ID int64
 }
